# Market.Api/Models/Product.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace Market.Api.Models
{
    public class Product
    {
        public int Id { get; set; }

        public string Name { get; set; }
    }
}